+++ eras/mary/impl/src/Cardano/Ledger/Mary/TxAuxData.hs
@@ -12,7 +12,7 @@ module Cardano.Ledger.Mary.TxAuxData (
 )
 where
 
-import Cardano.Ledger.Allegra.TxAuxData
+import Cardano.Ledger.Allegra.TxAuxData (AllegraTxAuxData (..), metadataAllegraTxAuxDataL)
 import Cardano.Ledger.AuxiliaryData (AuxiliaryDataHash (..))
 import Cardano.Ledger.Core (EraTxAuxData (..), upgradeScript)
 import Cardano.Ledger.Crypto (Crypto)
@@ -26,6 +26,10 @@ import Control.DeepSeq (deepseq)
 
 instance Crypto c => EraTxAuxData (MaryEra c) where
   type TxAuxData (MaryEra c) = AllegraTxAuxData (MaryEra c)
+
+  mkBasicTxAuxData = AllegraTxAuxData mempty mempty
+
+  metadataTxAuxDataL = metadataAllegraTxAuxDataL
 
   upgradeTxAuxData (AllegraTxAuxData md scripts) = AllegraTxAuxData md $ upgradeScript <$> scripts