+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/Tx/AuxData.hs
@@ -1,5 +1,7 @@
 module Cardano.Ledger.Api.Tx.AuxData (
   EraTxAuxData (TxAuxData),
+  mkBasicTxAuxData,
+  metadataTxAuxDataL,
   upgradeTxAuxData,
   hashTxAuxData,
   validateTxAuxData,